src/itron.meter.pkg/pyproject.toml

[project]
name = "itron.meter.pkg"
version = "1.0.0"
description = "Itron meter package"
authors = [
    {name = "Kurt Godwin", email = "kgodwin@itron.com"},
]
dependencies = [
    "GitPython>=3.1.27",
    "tblib>=1.7.0",
    "requests>=2.28.0",
    "hidapi>=0.13.1",
    "inquirer>=2.9.2",
    "PyYAML>=6.0",
    "paramiko>=2.11.0",
    "rpyc==5.3.0",
    "pyserial==3.5",
    "xmodem==0.4.6",
    "minimalmodbus==2.0.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "UNLICENSED"}

[project.scripts]
mm = "itron.scripts.mm:main"
mdb = "itron.scripts.mdb:main"
clean_locks = "itron.scripts.clean_locks:main"
ota-pack = "itron.scripts.otapack:main"
signer-client = "itron.scripts.signerclient:main"


[project.entry-points.pytest11]
parallel = "itron.plugins.xdist_meter"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
